requirements.txt
streamlit>=1.20
opencv-python-headless>=4.9.0
mediapipe>=0.10.0
numpy>=1.23
Pillow>=9.0
